Refactored TableRows component
Received onGetPosts, publishing and onPublishing props.

Received outlet context states.

Created handleUpdatePublish function to switch publish state of post.

Create a handleActiveModel function to pass DeletePostModel to the
model component, and then the model component will render the
DeletePostModel and handle deleting post.

The first of td uses dangerouslySetInnerHTML to render escaped post
title.

The button in the third td is replaced with link that navigate to
the editor.

Added the last td with loadIcon class.