Updated TableRows module css
Added the max-width with 300px value for tableRows class descendant
first of td tag in the media query for min width 900px viewport.

Added the max-width with 400px value for tableRows class descendant
first of td tag in the media query for min width 1200px viewport.

Added the max-width with 500px value for tableRows class descendant
first of td tag in the media query for min width 1600px viewport.